MB-65809 Add system:settings keyspace
Add a new system:settings keyspace for global settings for query. This
is where we intend to put all future query settings. The settings are
backed by metakv. Setting/changing of the settings are accomplished by
UPDATE statement on system:settings.

There is a single document in metakv for all settings, in
/query/settings/global_settings.

The settings are 2-level, 1st level is feature name, e.g.
plan_stability, 2nd level is the actual setting for plan_stability.

To set/update a setting, do:

  UPDATE system:settings SET plan_stability.enabled = true

note we use a 2-part field name plan_stability.enabled.

We can also do:

  UPDATE system:settings SET plan_stability = {"enabled": true}

although this is not recommended since it will discard any other
current settings for plan_stability (with only "enabled" remaining).
The previous version of the UPDATE statement will only update the
"enabled" setting and leaves any other potential settings untouched.
The 2nd version can be used when all settings for plan_stability is
set/updated at the same time.

Currently only "plan_stability" is supported, and the only available
setting is "enabled". This will be updated/enhanced as needed.

Change-Id: I01fa6ea026a0b1f91ca4229815644fc1a3e25f8b
Reviewed-on: https://review.couchbase.org/c/query/+/238085
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>